[cmyk-249] : home screen의 tab 수정, SpeakScreen 수정
