+++ lib/sample/cmyk-109/sample_chat_list_screen.dart
@@ -7,7 +7,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:intl/date_symbol_data_local.dart';
 
-import '../../screens/chat/ego_chat_room_list_screen.dart';
+import '../../screens/speak_screen.dart';
 
 /// AppBar 단위 테스트 코드
 /// SampleAppBarTest를 통해 위젯 비율을 조정하고 관리함
@@ -81,7 +81,7 @@ class _SampleMainAppBarScreenState extends ConsumerState<SampleMainScreen>
       body: TabBarView(
         controller: _tabController,
         children: [
-          TmpScreen(text: "스피크 페이지"), // TODO: 스피크 스크린에 연결
+          SpeakScreen(),
           RecordScreen(), // 캘린더 페이지
           ChatTabScreen()
         ],
